Move RuntimeConfig out of SVM
