+++ svm/src/lib.rs
@@ -7,7 +7,6 @@ pub mod message_processor;
 pub mod nonce_info;
 pub mod program_loader;
 pub mod rollback_accounts;
-pub mod runtime_config;
 pub mod transaction_account_state_info;
 pub mod transaction_commit_result;
 pub mod transaction_error_metrics;
